pyproject.toml
[tool.poetry]
name = "helpdesk-pet-project"
version = "0.1.0"
description = "A support system for a pet project"
authors = ["Yaroslav <yroslavkumpan@gmail.com>"]


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "5.2"
djangorestframework = "^3.16.1"
psycopg2 = "^2.9.11"
djangorestframework-simplejwt = "^5.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
